+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Added
 - Conductor like screening model (COSMO)
   implicit solvation model for SCC calculations
 
+- Printout of cavity information as a cosmo file
+
 - Extended syntax for selecting atoms in HSD input
 
 Changed